Add new function for seasonal max and min
For each year, end August to end August, annotate the min and max values
on the graph. At the same time, create a table of those mins/maxes
including the incident rate ratio comparing the 17/18 season to the 2022
season, with 95% confidence intervals.

The plot measures function is only used in one script, so it shouldn't
really be in the utilities file. Move that function, and non-general
helper functions into the plot_measures_report.py script.
