+++ analysis/report/report_utils.py
@@ -8,11 +8,8 @@ import seaborn as sns
 from collections import Counter
 from pathlib import Path
 from IPython.display import display, Markdown, HTML, Image
-import matplotlib.ticker as ticker
 
 colour_palette = sns.color_palette("Paired", 12)
-ticker.Locator.MAXTICKS = 10000
-
 
 SMALL_SIZE = 18
 MEDIUM_SIZE = 20
@@ -134,160 +131,7 @@ def parse_date(date_str):
 def add_date_lines(vlines, min_date, max_date):
     for date in vlines:
         if date >= min_date and date <= max_date:
-            try:
-                plt.axvline(x=date, color="orange", ls="--")
-            except parser._parser.ParserError:
-                continue
-
-
-def plot_measures(
-    df,
-    filename: str,
-    column_to_plot: str,
-    y_label: str,
-    as_bar: bool = False,
-    category: str = None,
-    frequency: str = "month",
-    log_scale: bool = False,
-    legend_inside: bool = False,
-    mark_seasons: bool = False,
-    date_lines: list = [],
-):
-    """Produce time series plot from measures table.  One line is plotted for each sub
-    category within the category column. Saves output as jpeg file.
-    Args:
-        df: A measure table
-        column_to_plot: Column name for y-axis values
-        y_label: Label to use for y-axis
-        as_bar: Boolean indicating if bar chart should be plotted instead of line chart. Only valid if no categories.
-        category: Name of column indicating different categories
-    """
-    df_copy = df.copy()
-
-    sns.set_style("darkgrid")
-    fig, ax = plt.subplots(figsize=(18, 10))
-
-    plt.rcParams["axes.prop_cycle"] = plt.cycler(color=colour_palette)
-    if log_scale:
-        plt.yscale("log")
-
-    # NOTE: finite filter for dummy data
-    y_max = (
-        df_copy[np.isfinite(df_copy[column_to_plot])][column_to_plot].max()
-        * 1.50
-    )
-    # Ignore timestamp - this could be done at load time
-    df_copy["date"] = df_copy["date"].dt.date
-
-    df_copy = df_copy.set_index("date")
-    if category:
-        # Set up category to have clean labels
-        repeated = autoselect_labels(df_copy[category])
-        df_copy[category] = df_copy.apply(
-            lambda x: translate_group(
-                x[category], x[category], repeated, True
-            ),
-            axis=1,
-        )
-        if as_bar:
-            df_copy.pivot(columns=category, values=column_to_plot).plot.bar(
-                ax=ax, stacked=True
-            )
-            y_max = (
-                df_copy.groupby(["date"])[column_to_plot].sum().max() * 1.05
-            )
-        else:
-            df_copy.groupby(category)[column_to_plot].plot(ax=ax)
-    else:
-        if as_bar:
-            df_copy[column_to_plot].bar(ax=ax, legend=False)
-        else:
-            df_copy[column_to_plot].plot(ax=ax, legend=False)
-
-    if frequency == "month":
-        xticks = pd.date_range(
-            start=df_copy.index.min(), end=df_copy.index.max(), freq="MS"
-        )
-        ax.set_xticks(xticks)
-        ax.set_xticklabels([x.strftime("%B %Y") for x in xticks])
-
-        if mark_seasons:
-            # TODO: check whether this works for weekly
-            df_copy["gas_year"] = pd.to_datetime(df_copy.index).to_period(
-                "A-Aug"
-            )
-            for year, data in df_copy.groupby(["gas_year", "name"]):
-                data_sorted = data.dropna(subset=[column_to_plot]).sort_values(
-                    column_to_plot
-                )
-                if data_sorted.empty:
-                    continue
-                data_min = data_sorted.iloc[0]
-                data_max = data_sorted.iloc[-1]
-                plt.annotate(
-                    int(data_min[column_to_plot]),
-                    xy=(data_min.name, data_min[column_to_plot]),
-                    xytext=(0, 25),
-                    textcoords="offset points",
-                    verticalalignment="bottom",
-                    arrowprops=dict(facecolor="black", shrink=0.025),
-                    horizontalalignment="left"
-                    if (data_min.name.month % 2) == 1
-                    else "right",
-                    fontsize=12,
-                )
-                plt.annotate(
-                    int(data_max[column_to_plot]),
-                    xy=(data_max.name, data_max[column_to_plot]),
-                    xytext=(0, -25),
-                    textcoords="offset points",
-                    verticalalignment="top",
-                    arrowprops=dict(facecolor="blue", shrink=0.025),
-                    fontsize=12,
-                )
-
-    elif frequency == "week":
-        xticks = pd.date_range(
-            start=df_copy.index.min(), end=df_copy.index.max(), freq="W-THU"
-        )
-        ax.set_xticks(xticks)
-        ax.set_xticklabels([x.strftime("%d-%m-%Y") for x in xticks])
-
-    if log_scale:
-        ax.yaxis.set_major_formatter(ticker.ScalarFormatter())
-        ax.yaxis.get_major_formatter().set_scientific(False)
-        y_label = f"Log {y_label.lower()}"
-
-    plt.ylabel(y_label)
-    plt.xlabel("Date")
-    plt.xticks(rotation="vertical")
-    plt.xticks(fontsize=14)
-
-    plt.ylim(
-        top=1000 if df_copy[column_to_plot].isnull().all() else y_max,
-    )
-
-    if category:
-        if legend_inside:
-            plt.legend(
-                sorted(df_copy[category].unique()),
-                ncol=3,
-            )
-        else:
-            plt.legend(
-                sorted(df_copy[category].unique()),
-                bbox_to_anchor=(1.04, 1),
-                loc="upper left",
-            )
-
-    if date_lines:
-        min_date = min(df_copy.index)
-        max_date = max(df_copy.index)
-        add_date_lines(date_lines, min_date, max_date)
-    plt.tight_layout()
-
-    plt.savefig(f"{filename}.jpeg")
-    plt.close()
+            plt.axvline(x=date, color="orange", ls="--")
 
 
 def coerce_numeric(table):
